配置数值读取增加上界校验,消除超范围配置值转窄整型的未定义行为 randrange 改在无符号域相减,消除 min/max 跨满 int32 区间时的有符号溢出 Lua core 网络接口取当前 task 后补空守卫,非 task 状态误用由崩溃转为 luaL_error mysql 二进制 float/double 列读取前置长度校验,畸形 server 数据由越界读转为返回失败 调试控制台存活列表为空时跳过 qsort,消除 qsort(NULL,0) 标准未定义行为

+++ lib/utils/utils.c
@@ -1228,7 +1228,7 @@ static uint64_t _xorshift64(void) {
 int32_t randrange(int32_t min, int32_t max) {
     // 闭区间 [min, max]：允许 max==min 退化情况返回 min；range==1 时 % 1 = 0 自然处理 
     ASSERTAB(max >= min, "rand range max must >= min.");
-    uint32_t range = (uint32_t)(max - min) + 1;
+    uint32_t range = ((uint32_t)max - (uint32_t)min) + 1;
     if (0 == range) {
         return min + (int32_t)(_xorshift64() >> 32);
     }
